Clean up, add test, fix sys mode
